✨ feat(ui): add Pressable component
The Pressable component is added to the UI components. It is a wrapper around the NativePressable component from react-native. It provides support for handling press events and applying a ripple effect on Android. The component also applies styles based on the pressed state.

📦 add(styles.js): add shadow style constant

The shadow style constant is added to the constants folder. It provides a consistent shadow style for both Android and iOS platforms. The style includes elevation for Android and shadow properties for iOS. The overflow property is also set based on the platform.